rename dep.py to requirements.py
