+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/dep.py
@@ -10,7 +10,7 @@ from .console import (
     CONTEXT_SETTINGS, abort, echo_failure, echo_info, echo_success, echo_waiting, echo_warning
 )
 from ..constants import get_root, get_agent_requirements
-from ..dep import (
+from ..requirements import (
     Package, make_catalog, read_packages, resolve_requirements
 )
 from ...utils import write_file_lines
