+++ datadog_checks_dev/datadog_checks/dev/tooling/commands/validate/agent_reqs.py
@@ -2,7 +2,6 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 import click
-import os
 
 from ..console import CONTEXT_SETTINGS, abort, echo_failure, echo_info, echo_success, echo_warning
 from ...constants import get_agent_release_requirements, AGENT_V5_ONLY, NOT_CHECKS
